scripts/requirements.txt
PyYAML>=6.0
paho-mqtt>=1.6.0
